+++ doc/index.rst
@@ -6,14 +6,12 @@
 Welcome to giottotime's API reference!
 ======================================
 
-Hello, how are you?
-
 .. include:: docs
 
-:mod:`giottotime.models`: Models
-====================================
+:mod:`giottotime.causality_tests`: Causality Tests
+====================================================
 
-.. automodule:: giottotime.models
+.. automodule:: giottotime.causality_tests
    :no-members:
    :no-inherited-members:
 
@@ -23,12 +21,13 @@ Hello, how are you?
    :toctree: generated/
    :template: class.rst
 
-   models.GAR
+   causality_tests.ShiftedLinearCorrelation
+   causality_tests.ShiftedPearsonCorrelation
 
-:mod:`giottotime.features.feature_creation`: Features Creation
-======================================================
+:mod:`giottotime.feature_creation`: Feature Creation
+====================================================
 
-.. automodule:: giottotime.features.feature_creation
+.. automodule:: giottotime.feature_creation
    :no-members:
    :no-inherited-members:
 
@@ -38,5 +37,28 @@ Hello, how are you?
    :toctree: generated/
    :template: class.rst
 
-   features.feature_creation.time_series_features.FeaturesCreation
-   features.feature_creation.time_series_features.ShiftFeature
+   feature_creation.FeaturesCreation
+   feature_creation.ShiftFeature
+   feature_creation.MovingAverageFeature
+   feature_creation.ConstantFeature
+   feature_creation.PolynomialFeature
+   feature_creation.ExogenousFeature
+   feature_creation.CustomFeature
+   feature_creation.CalendarFeature
+   feature_creation.PeriodicSeasonalFeature
+
+:mod:`giottotime.feature_creation.tda_features` TDA Features
+============================================================
+
+.. automodule:: giottotime.feature_creation.tda_features
+   :no-members:
+   :no-inherited-members:
+
+.. currentmodule:: giottotime
+
+.. autosummary::
+   :toctree: generated/
+   :template: class.rst
+
+   causality_tests.ShiftedLinearCorrelation
+   causality_tests.ShiftedPearsonCorrelation
