+++ giottotime/feature_creation/__init__.py
@@ -1,0 +1,16 @@
+from .calendar_features import CalendarFeature
+from .feature_creation import FeaturesCreation
+from .seasonal_features import PeriodicSeasonalFeature
+from .time_series_features import *
+
+__all__ = [
+    'FeaturesCreation',
+    'ShiftFeature',
+    'MovingAverageFeature',
+    'ConstantFeature',
+    'PolynomialFeature',
+    'ExogenousFeature',
+    'CustomFeature',
+    'CalendarFeature',
+    'PeriodicSeasonalFeature',
+]
