+++ giottotime/feature_creation/seasonal_features.py
