+++ giottotime/feature_creation/tda_features/average_lifetime_feature.py
@@ -77,7 +77,7 @@ class AvgLifeTimeFeature(TDAFeatures):
         indicating the distance between them.
 
     diags_homology_dimensions : ``Iterable``, optional, (default=``(0, 1)``)
-        Dimensions (non-negative integers) of the topological features_creation to be
+        Dimensions (non-negative integers) of the topological feature_creation to be
         detected.
 
     diags_coeff : ``int`` prime, optional, (default=``2``)
@@ -88,11 +88,11 @@ class AvgLifeTimeFeature(TDAFeatures):
     diags_max_edge_length : ``float``, optional, (default=``np.inf``)
         Upper bound on the maximum value of the Vietoris-Rips filtration
         parameter. Points whose distance is greater than this value will
-        never be connected by an edge, and topological features_creation at scales
+        never be connected by an edge, and topological feature_creation at scales
         larger than this value will not be detected.
 
     diags_infinity_values : ``float``, optional, (default=``None``)
-        Which death value to assign to features_creation which are still alive at
+        Which death value to assign to feature_creation which are still alive at
         filtration value `max_edge_length`. ``None`` has the same behaviour
         as `max_edge_length`.
 
@@ -147,7 +147,7 @@ class AvgLifeTimeFeature(TDAFeatures):
         Parameters
         ----------
         X : ``pd.DataFrame``, required.
-            The DataFrame on which to compute the features_creation.
+            The DataFrame on which to compute the feature_creation.
 
         Returns
         -------
