+++ giottotime/feature_creation/tda_features/relevant_holes_feature.py
@@ -1,58 +1,26 @@
-from typing import Iterable, Dict, Optional, Union, Callable
+from typing import Iterable, List, Optional, Union, Callable
+
+import numpy as np
+import pandas as pd
 
 from giottotime.features.tda_features.base import \
     TDAFeatures, _align_indices
 
-import pandas as pd
-import numpy as np
-import giotto.diagrams as diag
 
-
-class AmplitudeFeature(TDAFeatures):
+class NumberOfRelevantHolesFeature(TDAFeatures):
     """Compute the list of average lifetime for each time window, starting
     from the persistence diagrams.
 
     Parameters
     ----------
-    output_name: ``str``, required,
+    output_name : ``str``, required.
         The name of the output column
 
-    metric : ``'bottleneck'`` | ``'wasserstein'`` | ``'landscape'`` | \
-        ``'betti'`` | ``'heat'``, optional, (default=``'landscape'``)
-        Distance or dissimilarity function used to define the amplitude of
-        a subdiagram as its distance from the diagonal diagram:
-        - ``'bottleneck'`` and ``'wasserstein'`` refer to the identically named
-          perfect-matching--based notions of distance.
-        - ``'landscape'`` refers to the :math:`L^p` distance between
-          persistence landscapes.
-        - ``'betti'`` refers to the :math:`L^p` distance between Betti curves.
-        - ``'heat'`` refers to the :math:`L^p` distance between
-          Gaussian-smoothed diagrams.
+    h_dim: ``int``, optional, (default=``0``)
+        The homology dimension on which to compute the average lifetime.
 
-    amplitude_metric_params : ``Dict``, optional, (default=``None``)
-        Additional keyword arguments for the metric function:
-        - If ``metric == 'bottleneck'`` there are no available arguments.
-        - If ``metric == 'wasserstein'`` the only argument is `p` (int,
-          default: ``2``).
-        - If ``metric == 'betti'`` the available arguments are `p` (float,
-          default: ``2.``) and `n_values` (int, default: ``100``).
-        - If ``metric == 'landscape'`` the available arguments are `p`
-          (float, default: ``2.``), `n_values` (int, default: ``100``) and
-          `n_layers` (int, default: ``1``).
-        - If ``metric == 'heat'`` the available arguments are `p` (float,
-          default: ``2.``), `sigma` (float, default: ``1.``) and `n_values`
-          (int, default: ``100``).
-
-    amplitude_order : ``float``, optional, (default=``2.``)
-        If ``None``, :meth:`transform` returns for each diagram a vector of
-        amplitudes corresponding to the dimensions in
-        :attr:`homology_dimensions_`. Otherwise, the :math:`p`-norm of
-        these vectors with :math:`p` equal to `order` is taken.
-
-    amplitude_n_jobs : ``int``, optional, (default=``None``)
-        The number of jobs to use for the computation. ``None`` means 1 unless
-        in a :obj:`joblib.parallel_backend` context. ``-1`` means using all
-        processors.
+    theta: ``float``, optional, (default=``0.7``)
+        Constant used to set the threshold in the computation of the holes
 
     interpolation_strategy : ``str``, optional, (default=``ffill``)
         The interpolation strategy to use to fill the values
@@ -112,10 +80,10 @@ class AmplitudeFeature(TDAFeatures):
         indicating the distance between them.
 
     diags_homology_dimensions : ``Iterable``, optional, (default=``(0, 1)``)
-        Dimensions (non-negative integers) of the topological features_creation to be
+        Dimensions (non-negative integers) of the topological feature_creation to be
         detected.
 
-    diags_coeff : ``int`` prime, optional, (default=``2``)
+    diags_coeff : ``int`` (prime), optional, (default=``2``)
         Compute homology with coefficients in the prime field
         :math:`\\mathbb{F}_p = \\{ 0, \\ldots, p - 1 \\}` where
         :math:`p` equals `coeff`.
@@ -123,11 +91,11 @@ class AmplitudeFeature(TDAFeatures):
     diags_max_edge_length : ``float``, optional, (default=``np.inf``)
         Upper bound on the maximum value of the Vietoris-Rips filtration
         parameter. Points whose distance is greater than this value will
-        never be connected by an edge, and topological features_creation at scales
+        never be connected by an edge, and topological feature_creation at scales
         larger than this value will not be detected.
 
     diags_infinity_values : ``float``, optional, (default=``None``)
-        Which death value to assign to features_creation which are still alive at
+        Which death value to assign to feature_creation which are still alive at
         filtration value `max_edge_length`. ``None`` has the same behaviour
         as `max_edge_length`.
 
@@ -139,10 +107,8 @@ class AmplitudeFeature(TDAFeatures):
     """
     def __init__(self,
                  output_name: str,
-                 metric: str = 'landscape',
-                 amplitude_metric_params: Optional[Dict] = None,
-                 amplitude_order: Dict = 2,
-                 amplitude_n_jobs: Optional[float] = None,
+                 h_dim: int = 0,
+                 theta: float = 0.7,
                  interpolation_strategy: str = 'ffill',
                  takens_parameters_type: str = 'search',
                  takens_dimension: int = 5,
@@ -173,59 +139,66 @@ class AmplitudeFeature(TDAFeatures):
                          diags_infinity_values=diags_infinity_values,
                          diags_n_jobs=diags_n_jobs
                          )
-        self._metric = metric
-        self._amplitude_metric_params = amplitude_metric_params
-        self._amplitude_order = amplitude_order
-        self._amplitude_n_jobs = amplitude_n_jobs
+
+        self._h_dim = h_dim
+        self._theta = theta
         self._interpolation_strategy = interpolation_strategy
 
     def transform(self, X: pd.DataFrame) -> pd.DataFrame:
         """From the initial DataFrame ``X``, compute the persistence diagrams
-        and detect the average lifetime for a given homology dimension.
-        Then, assign a value to each initial data points, according to the
-        chosen ``interpolation_strategy``.
+        and detect the relevant number of holes. Then, assign a value to each
+        initial data points, according to the chosen ``interpolation_strategy``.
 
         Parameters
         ----------
         X : ``pd.DataFrame``, required.
-            The DataFrame on which to compute the features_creation.
+            The DataFrame on which to compute the feature_creation.
 
         Returns
         -------
         X_renamed : ``pd.DataFrame``
-            A DataFrame containing, for each original data-point, the average
-            lifetime associated to it. If, given the initial parameters, a
-            point was excluded from the computation, its value is set to
+            A DataFrame containing, for each original data-point, the relevant
+            number of holes associated to it. If, given the initial parameters,
+            a point wasexcluded from the computation, its value is set to
             ``Nan``.
 
         """
         persistence_diagrams = self._compute_persistence_diagrams(X)
-        amplitudes = self._calculate_amplitude_feature(persistence_diagrams)
-        original_points = self._compute_n_points(len(amplitudes))
+        n_holes = self._compute_num_relevant_holes(persistence_diagrams)
+        n_points = self._compute_n_points(len(n_holes))
 
-        X_aligned = _align_indices(X, original_points, amplitudes)
+        X_aligned = _align_indices(X, n_points, n_holes)
         X_renamed = self._rename_columns(X_aligned)
 
         return X_renamed
 
-    def _calculate_amplitude_feature(self, diagrams: np.ndarray) -> np.ndarray:
-        """Calculate the amplitude of the persistence diagrams
+    def _compute_num_relevant_holes(self, persistence_diagrams: np.ndarray)\
+            -> List:
+        """Compute the number of relevant holes in the point cloud.
 
         Parameters
         ----------
-        diagrams : ``np.ndarray``, required.
-            Array containing the persistence diagrams.
+        persistence_diagrams : ``np.ndarray``, required.
+            The array containing the scaled persistent diagrams.
 
         Returns
         -------
-        amplitudes : ``np.ndarray``
-            Array containing, for each diagrams, the corresponding amplitude,
-            calculated with with the given ``metric`` and ``amplitude_order``.
+        n_rel_holes : ``List``
+            For each diagram present in ``persistence_diagrams``, return the
+            number of relevant holes that have been found.
 
         """
+        n_rel_holes = []
+        for i in range(persistence_diagrams.shape[0]):
+            pers_table = pd.DataFrame(persistence_diagrams[i],
+                                      columns=['birth', 'death', 'homology'])
 
-        amplitude = diag.Amplitude(metric=self._metric,
-                                   order=self._amplitude_order,
-                                   metric_params=self._amplitude_metric_params,
-                                   n_jobs=self._amplitude_n_jobs)
-        return amplitude.fit_transform(diagrams)
+            pers_table['lifetime'] = pers_table['death'] - pers_table['birth']
+            threshold = pers_table[pers_table['homology'] == self._h_dim][
+                            'lifetime'].max() * self._theta
+            n_rel_holes.append(pers_table[
+                                   (pers_table['lifetime'] > threshold) & (
+                                           pers_table[
+                                               'homology'] == self._h_dim)].shape[0])
+
+        return n_rel_holes
