+++ giottotime/feature_creation/trend_features.py
