+++ giottotime/models/trend_models/custom_trend.py
